backend/requirements.txt
# Core dependencies for iReDev Framework
colorama>=0.4.4
pyyaml>=6.0
asyncio-mqtt>=0.11.0
aiofiles>=0.8.0

# Web interface dependencies
flask==3.0.3
flask-cors==4.0.1
pyjwt==2.8.0
python-dotenv==1.0.1
flask-sock==0.7.0

# AI integration dependencies
openai 
anthropic
google-generativeai
huggingface-hub
torch
tiktoken

# Langchain, Langgraph, and related dependencies
langchain
langchain-community
langchain-google-genai
langchain-huggingface
langchain-openai
langchain-postgres
langgraph
langgraph-checkpoint-postgres
langchain-text-splitters

# Other useful libraries
watchdog
semver
pypdf
psycopg[binary]
